+++ packages/sdk/CHANGELOG.md
@@ -13,6 +13,7 @@
   - Add `PendingValidatorScoreUpdate` to `UpdateInstructionPayload` union type.
   - Add `ChainParametersV3` to `ChainParameters` union type.
   - Add `isPrimedForSuspension` and `missedRounds` fields to `CurrentPaydayBakerPoolStatus`.
+  - Add suspended field to the `ConfigureBakerPayload`
 
 ## 8.1.1